move non Error specific helper methods to StringHelper
